+++ hata/discord/interaction/__init__.py
@@ -1,6 +1,7 @@
 from .application_command import *
 from .components import *
 from .event_types import *
+from .form import *
 from .preinstanced import *
 from .interaction_response_types import *
 from . import interaction_response_types as INTERACTION_RESPONSE_TYPES
@@ -10,6 +11,7 @@ __all__ = (
     *application_command.__all__,
     *components.__all__,
     *event_types.__all__,
+    *form.__all__,
     *interaction_response_types.__all__,
     *preinstanced.__all__,
 )